Keep scrolled down views through resizeEvents
There's a reason for everything...
ChatView is rendered into the viewport from top to bottom.
If we don't scroll to bottom, the topmost line doesn't change and therefore
the scrollbar will move upwards on shrinking events.
So if the scrollbar was at bottom before the resize, make it stay there.