Silence deprecation warnings from CMake 3.0+
Setting COMPILE_DEFINITIONS_<Config> has been deprecated since CMake 3.0.
However, the "proper" way of doing things now (generator expressions) have
been support only since CMake 2.8.10, and we require 2.8.1 currently.

So we explicitly tell CMake to use OLD behavior for CMP0043. And for CMP0026,
which is related to our handling of genversion (no need to fix this properly,
as genversion is gone in Quassel 0.11 anyway). And while we're at it, we
remove setting CMP0003 to NEW, this hasn't been needed in a very
long time...