Extend XDG_DATA_DIRS to include Quassel-specific directories
XDG_DATA_DIRS is used for example by Qt to find icon themes and other
application data. In order to avoid having to hack our own locations into
various places, we extend the variable at runtime to include them instead.

This is only relevant if KDE integration is not enabled; otherwise KDE will
specify a list of data dirs (this means that Quassel needs to be installed
properly for everything to work as expected if KDE integration is enabled).