Add command line parameter for overriding the system icon theme
Quassel uses many icons that are not part of the freedesktop naming
spec, but supported by KDE icon themes. If Quassel is run in
another environment, such as GNOME, which would select a non-KDE icon
theme, there may be missing icons.

In such cases, it is recommended to install the Oxygen icon theme and
start the Quassel client with '--icontheme=oxygen' to override the
system setting. Of course, you are free to choose any other theme as
well.

Note that oxygen is already selected as a fallback if the platform does
not tell Qt about its icon theme (for example, on non-Linux platforms).