Rename HAVE_KDE define to HAVE_KDE4
The build system now sets HAVE_KDE regardless if we're building
with KDE4 or KF5, and in addition adds HAVE_KDE4 and HAVE_KF5.
As many of the frameworks are source compatible with the old kdelibs,
this will allow switching HAVE_KDE4 back to HAVE_KDE later where
possible, without having to check for two defines everywhere.