Add support for Qt5's native command line parser
Qt 5.2 finally has grown a way for parsing command line arguments,
namely QCommandLineParser. According to the KF5 porting docs, this
is also the preferred choice for software using KDE Frameworks.

This commit adds a Qt5-based implementation for AbstractCliParser and
uses that instead of our own solution if Quassel is built against Qt5.
We also add support for naming the values that are shown in the --help
output, as QCommandLineParser can make use of that. Once we stop
supporting Qt4, the CLI parser abstraction shall be killed, simplifying
things a lot.