Bump Qt4 version requirement to 4.8
Qt 4.8 gained some support for C++11 features such as brace
initializers, which make some things much more convenient. As
this version was released more than three years ago, it should
be safe to finally bump our requirements accordingly. The core
was even fine with Qt 4.4 until now, and that was released a
whopping 7 years ago!

Requiring 4.8 will allow us to remove a bunch of old workarounds
as well, besides allowing for proper use of C++11 in the future.