Simplify session thread handling in Core
Whenever we want to get a session thread for a given user, we first
would check if we already have one and get that, or create a new one.
Instead of doing this several times in the code, this commit just
turns the former createSession() method into sessionForUser(), which
will create a thread if needed. This also removes a dead assignment
clang-analyzer complained about.