Add support for password hash versioning
Previously, Quassel would always use an unsalted SHA1 hash.  This
patch introduces a version field for the password, allowing for
using a better hash.  (No new hash is added yet.)  Users will be
automatically upgraded to the newest hash whenever they login or
change their passwords.

This is implemented by adding a new integer field to hold the hash
version in the DB.  Additionally, the password field in PostgreSQL
has been expanded to a TEXT field, allowing larger hash values to
be stored in the future.  The password hashing code has also been
refactored to make it easier to introduce new hashes without
fooling around in the DB driver code.