Work around weirdly broken MOCs
Apparently, Debian Wheezy ships Qt 4.8.2 with a MOC that thinks it's older
than 4.6, or something. Not sure what's going on, but if we remove
that "#if QT_VERSION" that checks for versions older than 4.6 (and thus does
not make any sense now that we require 4.8 anyway), the MOC on (some?)
Wheezy installations fails to deal with the Q_INTERFACES call.

I guess it doesn't hurt to keep the guard around, but this issue still leaves
me scratching my head...