Adapt custom stylesheet path to new default install location
When switching from a KDE4 build to Qt5/KF5, the default install
location for data files switches from /usr/share/apps/quassel to
/usr/share/quassel, as is standard. Users who have selected one of
the pre-installed stylesheets will use that setting.

Thus, if a custom stylesheet is configured and points to /usr/share/apps,
in case it is not found we will now try and look into /usr/share as well.
If we find the stylesheet in the new location, we'll also update the
stored setting.