Use KNotifications if Frameworks integration is enabled
This brings back proper support for KNotify if KDE Frameworks
integration is enabled.

Note that quassel.notifyrc is installed in a different location
if KF5 as opposed to KDE4. If you don't have a system installation
of Quassel, you may have to manually copy that file from data/
to $HOME/.local/share/knofitications5 for KNotify to know about
Quassel.